Draft to close #336.
