--- js/engine_serializer.js
@@ -1,228 +1,0 @@
-/**
- * Logic to serialize out snapshots of results from the engine.
- *
- * @license BSD-3-Clause
- */
-
-import {EngineNumber} from "engine_number";
-import {UnitConverter} from "engine_unit";
-import {OverridingConverterStateGetter} from "engine_unit_state";
-import {EngineResultBuilder, ImportSupplement} from "engine_struct";
-
-/**
- * Decorator around an engine to serialize out results.
- */
-class EngineResultSerializer {
-  /**
-   * Create a new decorator to produce engine result snapshots.
-   *
-   * @param {Engine} engine - The engine from which results will be serialized.
-   * @param {ConverterStateGetter} stateGetter - State getter by which to
-   *     access values at the current state that the engine is focusing on.
-   *     This will not be modified, only read.
-   */
-  constructor(engine, stateGetter) {
-    const self = this;
-    self._engine = engine;
-    self._stateGetter = stateGetter;
-  }
-
-  /**
-   * Serialize the results for an application and substance in a given year.
-   *
-   * @param {string} application - The name of the application for which a
-   *     result should be serialized like commerical refigeration.
-   * @param {string} substance - The name of the substance like HFC-134a for
-   *     which a result should be serialized.
-   * @param {number} year - The year for which a result should be serialized.
-   * @returns {EngineResult} - Snapshot of the result in the current engine
-   *     state for the given application and substance.
-   */
-  getResult(application, substance, year) {
-    const self = this;
-
-    // Create builder
-    const builder = new EngineResultBuilder();
-    builder.setApplication(application);
-    builder.setSubstance(substance);
-    builder.setYear(year);
-    builder.setScenarioName(self._engine.getScenarioName());
-    builder.setTrialNumber(self._engine.getTrialNumber());
-
-    // Add values into builder
-    self._parseMainBody(builder, application, substance);
-    self._parseImportSupplement(builder, application, substance);
-
-    // Will put build here instead of new EngineResult which is to be removed
-    return builder.build();
-  }
-
-  /**
-   * Parse the attributes which are actually returned to the user.
-   *
-   * @param {EngineResultBuilder} builder - The builder into which parsed
-   *     values should be registered.
-   * @param {string} application - The name of the application for which to get
-   *     values like commerical refrigeration.
-   * @param {string} substance - The name of the substance for which to get
-   *     values like HFC-134a.
-   */
-  _parseMainBody(builder, application, substance) {
-    const self = this;
-
-    // Prepare units
-    const stateGetter = new OverridingConverterStateGetter(self._stateGetter);
-    const unitConverter = new UnitConverter(stateGetter);
-
-    // Get sales
-    const manufactureRaw = self._engine.getStreamRaw(application, substance, "manufacture");
-    const importRaw = self._engine.getStreamRaw(application, substance, "import");
-    const recycleRaw = self._engine.getStreamRaw(application, substance, "recycle");
-
-    const manufactureValue = unitConverter.convert(manufactureRaw, "kg");
-    const importValue = unitConverter.convert(importRaw, "kg");
-    const recycleValue = unitConverter.convert(recycleRaw, "kg");
-    builder.setRecycleValue(recycleValue);
-
-    // Get total energy consumption
-    const energyConsumptionValue = self._engine.getStreamRaw(application, substance, "energy");
-    builder.setEnergyConsumption(energyConsumptionValue);
-
-    // Get emissions
-    const populationValue = self._engine.getStreamRaw(application, substance, "equipment");
-    builder.setPopulationValue(populationValue);
-
-    const populationNew = self._engine.getStreamRaw(application, substance, "newEquipment");
-    builder.setPopulationNew(populationNew);
-
-    const rechargeEmissions = self._engine.getStreamRaw(
-      application,
-      substance,
-      "rechargeEmissions",
-    );
-    const eolEmissions = self._engine.getStreamRaw(application, substance, "eolEmissions");
-    builder.setEolEmissions(eolEmissions);
-
-    // Get percent for offset
-    const manufactureKg = manufactureValue.getValue();
-    const importKg = importValue.getValue();
-    const recycleKg = recycleValue.getValue();
-
-    const nonRecycleSalesKg = manufactureKg + importKg;
-    const noSales = nonRecycleSalesKg == 0;
-    const percentManufacture = noSales ? 1 : manufactureKg / nonRecycleSalesKg;
-    const percentImport = 1 - percentManufacture;
-
-    // Offset sales
-    const manufactureValueOffset = new EngineNumber(
-      manufactureKg - recycleKg * percentManufacture,
-      "kg",
-    );
-    builder.setManufactureValue(manufactureValueOffset);
-
-    const importValueOffset = new EngineNumber(importKg - recycleKg * percentImport, "kg");
-    builder.setImportValue(importValueOffset);
-
-    // Get consumption
-    const getConsumptionByVolume = () => {
-      const consumptionRaw = self._engine.getGhgIntensity(application, substance);
-      const endsKg = consumptionRaw.getUnits().endsWith("kg");
-      const endsMt = consumptionRaw.getUnits().endsWith("mt");
-      if (endsKg || endsMt) {
-        return consumptionRaw;
-      } else {
-        return unitConverter.convert(consumptionRaw, "tCO2e / kg");
-      }
-    };
-    const consumptionByVolume = getConsumptionByVolume();
-
-    const getConsumptionForVolume = (volume) => {
-      if (volume.getValue() == 0) {
-        return new EngineNumber(0, "tCO2e");
-      }
-
-      stateGetter.setVolume(volume);
-      return unitConverter.convert(consumptionByVolume, "tCO2e");
-    };
-
-    const domesticConsumptionValue = getConsumptionForVolume(manufactureValueOffset);
-    builder.setDomesticConsumptionValue(domesticConsumptionValue);
-
-    const importConsumptionValue = getConsumptionForVolume(importValueOffset);
-    builder.setImportConsumptionValue(importConsumptionValue);
-
-    const recycleConsumptionValue = getConsumptionForVolume(recycleValue);
-    builder.setRecycleConsumptionValue(recycleConsumptionValue);
-
-    // Offset recharge emissions
-    stateGetter.setVolume(null);
-    const rechargeEmissionsConvert = unitConverter.convert(rechargeEmissions, "tCO2e");
-    const rechargeEmissionsOffset = new EngineNumber(
-      rechargeEmissionsConvert.getValue() - recycleConsumptionValue.getValue(),
-      "tCO2e",
-    );
-    builder.setRechargeEmissions(rechargeEmissionsOffset);
-  }
-
-  /**
-   * Parse information for the import supplement.
-   *
-   * Parse information for the import supplement which is needed to perform
-   * user-configurable import attribution options (are substances associated
-   * with importer or exporter).
-   *
-   * @param {EngineResultBuilder} builder - The builder into which parsed
-   *     values should be registered.
-   * @param {string} application - The name of the application for which to get
-   *     values like commerical refrigeration.
-   * @param {string} substance - The name of the substance for which to get
-   *     values like HFC-134a.
-   */
-  _parseImportSupplement(builder, application, substance) {
-    const self = this;
-
-    // Prepare units
-    const stateGetter = new OverridingConverterStateGetter(self._stateGetter);
-    const unitConverter = new UnitConverter(stateGetter);
-
-    const ghgIntensity = self._engine.getEqualsGhgIntensityFor(application, substance);
-    stateGetter.setSubstanceConsumption(ghgIntensity);
-
-    const importInitialChargeUnit = self._engine.getRawInitialChargeFor(
-      application,
-      substance,
-      "import",
-    );
-    stateGetter.setAmortizedUnitVolume(importInitialChargeUnit);
-
-    // Determine import value without recharge
-    const totalImportValue = self._engine.getStreamRaw(application, substance, "import");
-    const totalDomesticValue = self._engine.getStreamRaw(application, substance, "manufacture");
-    const totalRechargeEmissions = self._engine.getStreamRaw(
-      application,
-      substance,
-      "rechargeEmissions",
-    );
-
-    const totalImportValueKg = unitConverter.convert(totalImportValue, "kg").getValue();
-    const totalDomesticValueKg = unitConverter.convert(totalDomesticValue, "kg").getValue();
-    const totalKg = totalImportValueKg + totalDomesticValueKg;
-    const proportionImport = totalKg == 0 ? 0 : totalImportValueKg / totalKg;
-    const totalRechargeKg = unitConverter.convert(totalRechargeEmissions, "kg").getValue();
-
-    const importRechargeKg = proportionImport * totalRechargeKg;
-    const importForInitialChargeKg = totalImportValueKg - importRechargeKg;
-
-    const value = new EngineNumber(importForInitialChargeKg, "kg");
-
-    // Determine consumption (tCO2e) and population (units)
-    const consumption = unitConverter.convert(value, "tCO2e");
-    const population = unitConverter.convert(value, "units");
-
-    // Package
-    const importSupplement = new ImportSupplement(value, consumption, population);
-    builder.setImportSupplement(importSupplement);
-  }
-}
-
-export {EngineResultSerializer};
