--- js/legacy_backend.js
@@ -1,331 +1,0 @@
-/**
- * Legacy JavaScript backend for executing QubecTalk simulations via web worker.
- *
- * @license BSD, see LICENSE.md.
- */
-
-import {EngineNumber} from "engine_number";
-import {EngineResult} from "engine_struct";
-
-/**
- * Parser for handling CSV report data returned from the legacy JS worker.
- */
-class ReportDataParser {
-  /**
-   * Parse the response from the legacy worker.
-   *
-   * @param {string} response - The response string from worker containing status and CSV data.
-   * @returns {Array<EngineResult>} Parsed engine results.
-   * @throws {Error} If response indicates an error status.
-   */
-  static parseResponse(response) {
-    const lines = response.split("\n");
-
-    if (lines.length < 2) {
-      throw new Error("Invalid response format: missing status line or data");
-    }
-
-    const status = lines[0].trim();
-    if (status !== "OK") {
-      throw new Error(status);
-    }
-
-    // Skip empty line after status
-    const csvData = lines.slice(2).join("\n").trim();
-
-    if (!csvData) {
-      return [];
-    }
-
-    // Check if we have only headers (for testing/placeholder mode)
-    const csvLines = csvData.split("\n").filter((line) => line.trim());
-    if (csvLines.length <= 1) {
-      return []; // Only headers, no data
-    }
-
-    return ReportDataParser._parseCsvData(csvData);
-  }
-
-  /**
-   * Parse CSV data into EngineResult objects.
-   *
-   * @private
-   * @param {string} csvData - The CSV data to parse.
-   * @returns {Array<EngineResult>} Array of parsed engine results.
-   */
-  static _parseCsvData(csvData) {
-    const lines = csvData.split("\n").filter((line) => line.trim());
-
-    if (lines.length === 0) {
-      return [];
-    }
-
-    // Parse header to understand column structure
-    const headers = lines[0].split(",").map((h) => h.trim());
-    const results = [];
-
-    for (let i = 1; i < lines.length; i++) {
-      const values = lines[i].split(",").map((v) => v.trim());
-
-      if (values.length !== headers.length) {
-        continue; // Skip malformed rows
-      }
-
-      const row = {};
-      headers.forEach((header, index) => {
-        row[header] = values[index];
-      });
-
-      try {
-        const engineResult = ReportDataParser._createEngineResult(row);
-        results.push(engineResult);
-      } catch (e) {
-        console.warn("Failed to parse row:", row, e);
-        // Continue parsing other rows
-      }
-    }
-
-    return results;
-  }
-
-  /**
-   * Create an EngineResult from a parsed CSV row.
-   *
-   * @private
-   * @param {Object} row - The parsed CSV row data.
-   * @returns {EngineResult} The created engine result.
-   */
-  static _createEngineResult(row) {
-    // Helper function to create EngineNumber from row data
-    const getEngineNumber = (valueKey, unitsKey, defaultUnits = "units") => {
-      const value = parseFloat(row[valueKey] || "0");
-      const units = row[unitsKey] || defaultUnits;
-      return new EngineNumber(value, units);
-    };
-
-    const application = row["application"] || "";
-    const substance = row["substance"] || "";
-    const year = parseInt(row["year"] || "0");
-    const scenarioName = row["scenarioName"] || "";
-    const trialNumber = parseInt(row["trialNumber"] || "0");
-
-    // Create EngineNumber objects for all the numeric fields
-    const manufactureValue = getEngineNumber("manufacture", "manufactureUnits", "kg");
-    const importValue = getEngineNumber("import", "importUnits", "kg");
-    const recycleValue = getEngineNumber("recycle", "recycleUnits", "kg");
-    const domesticConsumptionValue = getEngineNumber(
-      "domesticConsumption",
-      "domesticConsumptionUnits",
-      "tCO2e",
-    );
-    const importConsumptionValue = getEngineNumber(
-      "importConsumption",
-      "importConsumptionUnits",
-      "tCO2e",
-    );
-    const recycleConsumptionValue = getEngineNumber(
-      "recycleConsumption",
-      "recycleConsumptionUnits",
-      "tCO2e",
-    );
-    const populationValue = getEngineNumber("population", "populationUnits", "units");
-    const populationNew = getEngineNumber("populationNew", "populationNewUnits", "units");
-    const rechargeEmissions = getEngineNumber(
-      "rechargeEmissions",
-      "rechargeEmissionsUnits",
-      "tCO2e",
-    );
-    const eolEmissions = getEngineNumber("eolEmissions", "eolEmissionsUnits", "tCO2e");
-    const energyConsumption = getEngineNumber("energyConsumption", "energyConsumptionUnits", "kwh");
-
-    return new EngineResult(
-      application,
-      substance,
-      year,
-      scenarioName,
-      trialNumber,
-      manufactureValue,
-      importValue,
-      recycleValue,
-      domesticConsumptionValue,
-      importConsumptionValue,
-      recycleConsumptionValue,
-      populationValue,
-      populationNew,
-      rechargeEmissions,
-      eolEmissions,
-      energyConsumption,
-      null, // importSupplement - not used in legacy backend
-    );
-  }
-}
-
-/**
- * Web worker layer for managing communication with the legacy JS execution worker.
- */
-class LegacyJsLayer {
-  /**
-   * Create a new LegacyJsLayer instance.
-   */
-  constructor() {
-    const self = this;
-    self._worker = null;
-    self._initPromise = null;
-    self._pendingRequests = new Map();
-    self._nextRequestId = 1;
-  }
-
-  /**
-   * Initialize the worker and prepare for execution.
-   *
-   * @returns {Promise<void>} Promise that resolves when worker is ready.
-   */
-  initialize() {
-    const self = this;
-
-    if (self._initPromise !== null) {
-      return self._initPromise;
-    }
-
-    self._initPromise = new Promise((resolve, reject) => {
-      try {
-        self._worker = new Worker("/js/legacy.worker.js");
-
-        self._worker.onmessage = (event) => {
-          self._handleWorkerMessage(event);
-        };
-
-        self._worker.onerror = (error) => {
-          console.error("Worker error:", error);
-          reject(new Error("Worker failed to load: " + error.message));
-        };
-
-        // Worker is ready immediately for legacy JS
-        resolve();
-      } catch (error) {
-        reject(error);
-      }
-    });
-
-    return self._initPromise;
-  }
-
-  /**
-   * Execute QubecTalk code and return parsed results.
-   *
-   * @param {string} code - The QubecTalk code to execute.
-   * @returns {Promise<Array<EngineResult>>} Promise resolving to execution results.
-   */
-  async runSimulation(code) {
-    const self = this;
-
-    await self.initialize();
-
-    return new Promise((resolve, reject) => {
-      const requestId = self._nextRequestId++;
-
-      self._pendingRequests.set(requestId, {resolve, reject});
-
-      self._worker.postMessage({
-        id: requestId,
-        command: "execute",
-        code: code,
-      });
-
-      // Set timeout for long-running simulations
-      setTimeout(() => {
-        if (self._pendingRequests.has(requestId)) {
-          self._pendingRequests.delete(requestId);
-          reject(new Error("Simulation timeout"));
-        }
-      }, 30000); // 30 second timeout
-    });
-  }
-
-  /**
-   * Handle messages from the worker.
-   *
-   * @private
-   * @param {MessageEvent} event - The message event from worker.
-   */
-  _handleWorkerMessage(event) {
-    const self = this;
-    const {id, success, result, error} = event.data;
-
-    const request = self._pendingRequests.get(id);
-    if (!request) {
-      console.warn("Received response for unknown request:", id);
-      return;
-    }
-
-    self._pendingRequests.delete(id);
-
-    if (success) {
-      try {
-        const parsedResults = ReportDataParser.parseResponse(result);
-        request.resolve(parsedResults);
-      } catch (parseError) {
-        request.reject(parseError);
-      }
-    } else {
-      request.reject(new Error(error || "Unknown worker error"));
-    }
-  }
-
-  /**
-   * Terminate the worker and clean up resources.
-   */
-  terminate() {
-    const self = this;
-
-    if (self._worker) {
-      self._worker.terminate();
-      self._worker = null;
-    }
-
-    // Reject any pending requests
-    for (const [id, request] of self._pendingRequests) {
-      request.reject(new Error("Worker terminated"));
-    }
-    self._pendingRequests.clear();
-
-    self._initPromise = null;
-  }
-}
-
-/**
- * Legacy JavaScript backend for executing QubecTalk simulations.
- *
- * This backend executes QubecTalk code in a web worker for isolation
- * and returns the results as parsed EngineResult objects.
- */
-class LegacyJsBackend {
-  /**
-   * Create a new LegacyJsBackend instance.
-   *
-   * @param {LegacyJsLayer} legacyJsLayer - The layer for worker communication.
-   */
-  constructor(legacyJsLayer) {
-    const self = this;
-    self._legacyJsLayer = legacyJsLayer;
-  }
-
-  /**
-   * Execute QubecTalk simulation code.
-   *
-   * @param {string} simCode - The QubecTalk code to execute.
-   * @returns {Promise<Array<EngineResult>>} Promise resolving to simulation results.
-   */
-  async execute(simCode) {
-    const self = this;
-
-    try {
-      const results = await self._legacyJsLayer.runSimulation(simCode);
-      return results;
-    } catch (error) {
-      throw new Error("Simulation execution failed: " + error.message);
-    }
-  }
-}
-
-export {LegacyJsBackend, LegacyJsLayer, ReportDataParser};
