+++ js/ui_editor.js
@@ -4,8 +4,7 @@
  * @license BSD, see LICENSE.md.
  */
 import {EngineNumber} from "engine_number";
-
-import {YearMatcher} from "engine_state";
+import {YearMatcher} from "year_matcher";
 
 import {
   Application,
