+++ js/ui_translator.js
@@ -8,7 +8,7 @@
  */
 
 import {EngineNumber} from "engine_number";
-import {YearMatcher} from "engine_state";
+import {YearMatcher} from "year_matcher";
 
 /**
  * Command compatibility mapping to compatibility modes:
