+++ test/test.html
@@ -25,25 +25,15 @@
                 "compiler": "../js/compiler.js?v=EPOCH",
                 "engine_const": "../js/engine_const.js?v=EPOCH",
                 "engine_number": "../js/engine_number.js?v=EPOCH",
-                "engine_unit": "../js/engine_unit.js?v=EPOCH",
-                "engine_unit_state": "../js/engine_unit_state.js?v=EPOCH",
-                "engine_serializer": "../js/engine_serializer.js?v=EPOCH",
-                "engine_state": "../js/engine_state.js?v=EPOCH",
                 "engine_struct": "../js/engine_struct.js?v=EPOCH",
-                "engine": "../js/engine.js?v=EPOCH",
-                "legacy_backend": "../js/legacy_backend.js?v=EPOCH",
+                "year_matcher": "../js/year_matcher.js?v=EPOCH",
                 "wasm_backend": "../js/wasm_backend.js?v=EPOCH",
                 "report_data": "../js/report_data.js?v=EPOCH",
                 "ui_translator": "../js/ui_translator.js?v=EPOCH",
                 "user_config": "../js/user_config.js?v=EPOCH",
-                "test_compiler": "./test_compiler.js?v=EPOCH",
+                "test_integration": "./test_integration.js?v=EPOCH",
                 "test_engine_number": "./test_engine_number.js?v=EPOCH",
-                "test_engine_unit_converter": "./test_engine_unit_converter.js?v=EPOCH",
-                "test_engine_unit_state": "./test_engine_unit_state.js?v=EPOCH",
-                "test_engine_serializer": "./test_engine_serializer.js?v=EPOCH",
-                "test_engine_state": "./test_engine_state.js?v=EPOCH",
                 "test_engine_struct": "./test_engine_struct.js?v=EPOCH",
-                "test_engine": "./test_engine.js?v=EPOCH",
                 "test_report_data": "./test_report_data.js?v=EPOCH",
                 "test_ui_translator": "./test_ui_translator.js?v=EPOCH",
                 "test_ui_translator_reverse": "./test_ui_translator_reverse.js?v=EPOCH",
@@ -59,31 +49,21 @@
     </script>
 
     <script type="module">
-        import {buildCompilerTests} from "test_compiler";
+        import {buildCompilerTests} from "test_integration";
         import {buildEngineNumberTests} from "test_engine_number";
-        import {buildEngineUnitConverterTests} from "test_engine_unit_converter";
-        import {buildEngineUnitStateTests} from "test_engine_unit_state";
-        import {buildEngineStateTests} from "test_engine_state";
         import {buildEngineStructTests} from "test_engine_struct";
-        import {buildEngineTests} from "test_engine";
         import {buildReportDataTests} from "test_report_data";
         import {buildUiTranslatorTests} from "test_ui_translator";
         import {buildUiTranslatorReverseTests} from "test_ui_translator_reverse";
-        import {buildEngineSerializerTests} from "test_engine_serializer";
         import {buildMainTests} from "test_main";
         import {buildWasmBackendTests} from "test_wasm_backend";
 
         buildCompilerTests();
         buildEngineNumberTests();
-        buildEngineUnitConverterTests();
-        buildEngineUnitStateTests();
-        buildEngineStateTests();
         buildEngineStructTests();
-        buildEngineTests();
         buildReportDataTests();
         buildUiTranslatorTests();
         buildUiTranslatorReverseTests();
-        buildEngineSerializerTests();
         buildMainTests();
         buildWasmBackendTests();
     </script>
