+++ test/test_ui_translator_reverse.js
@@ -1,5 +1,5 @@
 import {EngineNumber} from "engine_number";
-import {YearMatcher} from "engine_state";
+import {YearMatcher} from "year_matcher";
 import {
   AboutStanza,
   Application,
